# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BLACK Trading evaluation, checkout, payment, and administration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Challenge {
  id: string;
  displayName: string;
  accountSize: string;
  price: number;
  description: string;
  rulesSummary: string;
  enabled: boolean;
  launchOffer: boolean;
  sortOrder: number;
}

export interface ChallengeInput {
  /** @minLength 1 */
  displayName: string;
  /** @minLength 1 */
  accountSize: string;
  /** @minimum 0 */
  price: number;
  description: string;
  rulesSummary: string;
  enabled: boolean;
  launchOffer: boolean;
  sortOrder: number;
}

export interface Faq {
  id: string;
  question: string;
  answer: string;
  sortOrder: number;
  published: boolean;
}

export interface FaqInput {
  /** @minLength 1 */
  question: string;
  answer: string;
  sortOrder: number;
  published: boolean;
}

export interface Testimonial {
  id: string;
  customerName: string;
  quote: string;
  verified: boolean;
  published: boolean;
  demo: boolean;
}

export interface TestimonialInput {
  /** @minLength 1 */
  customerName: string;
  quote: string;
  verified: boolean;
  published: boolean;
  demo: boolean;
}

export type MediaItemKind = typeof MediaItemKind[keyof typeof MediaItemKind];


export const MediaItemKind = {
  video: 'video',
  screenshot: 'screenshot',
  story: 'story',
  announcement: 'announcement',
} as const;

export interface MediaItem {
  id: string;
  title: string;
  kind: MediaItemKind;
  caption: string;
  /** @nullable */
  assetUrl: string | null;
  published: boolean;
  demo: boolean;
  sortOrder: number;
}

export type MediaInputKind = typeof MediaInputKind[keyof typeof MediaInputKind];


export const MediaInputKind = {
  video: 'video',
  screenshot: 'screenshot',
  story: 'story',
  announcement: 'announcement',
} as const;

export interface MediaInput {
  /** @minLength 1 */
  title: string;
  kind: MediaInputKind;
  caption: string;
  /** @nullable */
  assetUrl?: string | null;
  published: boolean;
  demo: boolean;
  sortOrder: number;
}

export interface SiteStatistic {
  id: string;
  label: string;
  value: string;
  verified: boolean;
  visible: boolean;
  sortOrder: number;
}

export type StatisticsInputStatisticsItem = {
  label: string;
  value: string;
  verified: boolean;
  visible: boolean;
  sortOrder: number;
};

export interface StatisticsInput {
  statistics: StatisticsInputStatisticsItem[];
}

export type SiteSettingsSocialLinksItem = {
  label: string;
  url: string;
};

export interface SiteSettings {
  companyName: string;
  logoText: string;
  supportEmail: string;
  socialLinks: SiteSettingsSocialLinksItem[];
  termsContent: string;
  privacyContent: string;
}

export type SiteSettingsInputSocialLinksItem = {
  label: string;
  url: string;
};

export interface SiteSettingsInput {
  companyName: string;
  logoText: string;
  supportEmail: string;
  socialLinks: SiteSettingsInputSocialLinksItem[];
  termsContent: string;
  privacyContent: string;
}

export interface PaymentMethod {
  id: string;
  currency: string;
  network: string;
  walletAddress: string;
  enabled: boolean;
  instructions: string;
  updatedAt: string;
}

export interface PaymentMethodInput {
  /** @minLength 1 */
  currency: string;
  /** @minLength 1 */
  network: string;
  /** @minLength 1 */
  walletAddress: string;
  enabled: boolean;
  instructions: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  PENDING: 'PENDING',
  PAYMENT_SUBMITTED: 'PAYMENT_SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  PAID: 'PAID',
  REJECTED: 'REJECTED',
  REFUNDED: 'REFUNDED',
  CANCELLED: 'CANCELLED',
} as const;

export interface Order {
  id: string;
  fullName: string;
  email: string;
  country: string;
  /** @nullable */
  phone: string | null;
  challengeId: string;
  challengeName: string;
  accountSize: string;
  price: number;
  /** @nullable */
  paymentMethodId: string | null;
  /** @nullable */
  cryptoAmount: number | null;
  /** @nullable */
  currency: string | null;
  /** @nullable */
  network: string | null;
  /** @nullable */
  paymentAddress: string | null;
  /** @nullable */
  transactionHash: string | null;
  paymentStatus: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  fullName: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /**
     * @maxLength 40
     * @nullable
     */
  phone?: string | null;
  challengeId: string;
}

export interface PaymentQuote {
  paymentMethodId: string;
  currency: string;
  network: string;
  walletAddress: string;
  instructions: string;
  amount: number;
  quoteUpdatedAt: string;
  quoteSource: string;
  quoteToken: string;
}

export interface OrderPayment {
  id: string;
  challengeName: string;
  accountSize: string;
  price: number;
  paymentStatus: OrderStatus;
  /** @nullable */
  currency: string | null;
  /** @nullable */
  network: string | null;
  /** @nullable */
  paymentAddress: string | null;
  /** @nullable */
  instructions: string | null;
  paymentOptions: PaymentQuote[];
}

export interface PaymentSubmissionInput {
  /**
     * @minLength 2
     * @maxLength 12
     */
  currency: string;
  /**
     * @minLength 2
     * @maxLength 40
     */
  network: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  transactionHash: string;
  /**
     * @minLength 32
     * @maxLength 4096
     */
  quoteToken: string;
}

export interface OrderStatusUpdate {
  paymentStatus: OrderStatus;
}

export type PublicSiteSocialLinksItem = {
  label: string;
  url: string;
};

export interface PublicSite {
  companyName: string;
  logoText: string;
  supportEmail: string;
  socialLinks: PublicSiteSocialLinksItem[];
  termsContent: string;
  privacyContent: string;
  challenges: Challenge[];
  faqs: Faq[];
  testimonials: Testimonial[];
  media: MediaItem[];
  statistics: SiteStatistic[];
}

export interface AdminDashboard {
  totalOrders: number;
  pendingPayments: number;
  paidOrders: number;
  rejectedOrders: number;
  revenue: number;
  recentOrders: Order[];
}