# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BLACK Trading evaluation, checkout, payment, and administration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a pending order using the server-side challenge price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Pending order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid customer information
        "404":
          description: Challenge is unavailable
  /orders/{orderId}/payment:
    get:
      operationId: getOrderPayment
      tags: [orders]
      summary: Get limited payment-page details for an order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order details safe for the payment page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPayment"
        "404":
          description: Order not found
    post:
      operationId: submitPayment
      tags: [orders]
      summary: Submit a transaction hash for manual review
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSubmissionInput"
      responses:
        "200":
          description: Payment submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPayment"
        "400":
          description: Invalid payment submission
        "404":
          description: Order or payment method not found
...
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    OrderStatus:
      type: string
      enum: [PENDING, PAYMENT_SUBMITTED, UNDER_REVIEW, PAID, REJECTED, REFUNDED, CANCELLED]
    Order:
      type: object
      required: [id, fullName, email, country, phone, challengeId, challengeName, accountSize, price, paymentMethodId, cryptoAmount, currency, network, paymentAddress, transactionHash, paymentStatus, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        email: { type: string, format: email }
        country: { type: string }
        phone: { type: ["string", "null"] }
        challengeId: { type: string, format: uuid }
        challengeName: { type: string }
        accountSize: { type: string }
        price: { type: number }
        paymentMethodId: { type: ["string", "null"], format: uuid }
        cryptoAmount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        network: { type: ["string", "null"] }
        paymentAddress: { type: ["string", "null"] }
        transactionHash: { type: ["string", "null"] }
        paymentStatus: { $ref: "#/components/schemas/OrderStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrderInput:
      type: object
      required: [fullName, email, country, challengeId]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 160 }
        email: { type: string, format: email, maxLength: 320 }
        country: { type: string, minLength: 2, maxLength: 100 }
        phone: { type: ["string", "null"], maxLength: 40 }
        challengeId: { type: string, format: uuid }
    OrderPayment:
      type: object
      required: [id, challengeName, accountSize, price, paymentStatus, currency, network, paymentAddress, instructions, paymentOptions]
      properties:
        id: { type: string, format: uuid }
        challengeName: { type: string }
        accountSize: { type: string }
        price: { type: number }
        paymentStatus: { $ref: "#/components/schemas/OrderStatus" }
        currency: { type: ["string", "null"] }
        network: { type: ["string", "null"] }
        paymentAddress: { type: ["string", "null"] }
        instructions: { type: ["string", "null"] }
        paymentOptions:
          type: array
          items: { $ref: "#/components/schemas/PaymentQuote" }
    PaymentQuote:
      type: object
      required: [paymentMethodId, currency, network, walletAddress, instructions, amount, quoteUpdatedAt, quoteSource, quoteToken]
      properties:
        paymentMethodId: { type: string, format: uuid }
        currency: { type: string }
        network: { type: string }
        walletAddress: { type: string }
        instructions: { type: string }
        amount: { type: number }
        quoteUpdatedAt: { type: string, format: date-time }
        quoteSource: { type: string }
        quoteToken: { type: string }
    PaymentSubmissionInput:
      type: object
      required: [currency, network, transactionHash, quoteToken]
      properties:
        currency: { type: string, minLength: 2, maxLength: 12 }
        network: { type: string, minLength: 2, maxLength: 40 }
        transactionHash: { type: string, minLength: 8, maxLength: 200 }
        quoteToken: { type: string, minLength: 32, maxLength: 4096 }
...