# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BLACK Trading evaluation, checkout, payment, and administration API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  fullName: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /**
     * @maxLength 40
     * @nullable
     */
  phone?: string | null;
  challengeId: string;
}